Rename file from blog to views
